f64 is overkill

+++ src/automasking.rs
@@ -2,10 +2,10 @@ use log::info;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-pub type MaskThresholdMap = HashMap<String, f64>;
+pub type MaskThresholdMap = HashMap<String, f32>;
 
 pub struct AutoMaskSampler {
-    threshold_margin: f64,
+    threshold_margin: f32,
     pub angles_with_thresholds: MaskThresholdMap,
     scans_remaining: usize,
 }
@@ -16,7 +16,7 @@ pub struct AutoMaskMessage {
 }
 
 impl AutoMaskSampler {
-    pub fn new(required_scans_count: usize, threshold_margin: f64) -> AutoMaskSampler {
+    pub fn new(required_scans_count: usize, threshold_margin: f32) -> AutoMaskSampler {
         AutoMaskSampler {
             threshold_margin,
             angles_with_thresholds: HashMap::new(),
@@ -27,7 +27,7 @@ impl AutoMaskSampler {
     /** Add samples (vector of angles with distances) until sufficient scans have been recorded;
      * return the mapping once we're done, otherwise return None
      */
-    pub fn add_samples(&mut self, samples: &Vec<(f64, f64)>) -> Option<&MaskThresholdMap> {
+    pub fn add_samples(&mut self, samples: &Vec<(f32, f32)>) -> Option<&MaskThresholdMap> {
         self.scans_remaining -= 1;
 
         if self.scans_remaining > 0 {
